requirements.txt
Flask==2.3.3
requests==2.31.0
Pillow==10.0.1